conflictedit

+++ src/group9/Booking.java
@@ -7,10 +7,10 @@ public class Booking {
     int roomNo;
     long checkInDate;
     long checkOutDate;
-    float price;
+    double price;
 
 
-    public Booking(int roomNo, long checkInDate, long checkOutDate, float price) {
+    public Booking(int roomNo, long checkInDate, long checkOutDate, double price) {
         this.roomNo = roomNo;
         this.checkInDate = checkInDate;
         this.checkOutDate = checkOutDate;
@@ -43,7 +43,7 @@ public class Booking {
     }
 
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
